fix: double-click paste bypasses selection index and gesture arbitration
Double-clicking a card to paste it silently no-op'd when the click
landed on a tappable chip (v3.3.0): ChipRow's chip tap attaches via
.simultaneousGesture on a descendant view, which is exempt from
SwiftUI's cross-view gesture arbitration, so the card's ancestor
.onTapGesture(count: 2) could lose the race outright. Separately,
that gesture resolved what to paste through state.selectedIndex
rather than the card actually clicked — a second layer of risk if the
two ever disagreed.

Fixes both: the card's double-click now uses .highPriorityGesture so
it wins arbitration unconditionally regardless of descendant gestures,
and it pastes the clicked entry's id directly via a new
PopupController.paste(entryId:) path (PasteAction/PasteboardWriter now
take an injectable pasteboard for test isolation). pasteSelected()
keeps reading selectedIndex for the Return-key path, where that's
legitimate.

Adds a "paste" os.Logger category at .log/.error/.warning (not .info,
which never survives to `log show`) at every step of the flow —
gesture fire, entry point, PasteboardWriter's previously-discarded
writeObjects Bool, NSRunningApplication.activate()'s discarded Bool,
and synthesizeCmdV's silent nil-CGEvent no-op — logging numeric entry
ids only, never content.

+++ Sources/CpdbApp/Popup/PopupRootView.swift
@@ -5,7 +5,10 @@ import CpdbShared
 /// Top-level SwiftUI view hosted inside `PopupPanel`.
 struct PopupRootView: View {
     @Bindable var state: PopupState
-    let onPaste: () -> Void
+    /// Threaded straight down to `EntryStripView`'s double-click gesture —
+    /// see that type's doc comment for why it takes an id rather than
+    /// just re-pasting the current selection.
+    let onPaste: (Int64) -> Void
 
     var body: some View {
         VStack(alignment: .leading, spacing: 0) {
